+++ packages/map-template/src/components/MapsIndoorsMap/MapsIndoorsMap.jsx
@@ -179,7 +179,9 @@ function MapsIndoorsMap(props) {
 
     return (
         <RecoilRoot>
-            {mapTemplateProps && <MapTemplate {...mapTemplateProps}></MapTemplate>}
+            <Sentry.ErrorBoundary>
+                {mapTemplateProps && <MapTemplate {...mapTemplateProps}></MapTemplate>}
+            </Sentry.ErrorBoundary>
         </RecoilRoot>
     )
 }
@@ -216,4 +218,4 @@ Sentry.init({
     replaysOnErrorSampleRate: 1.0,
 });
 
-export default MapsIndoorsMap;
+export default Sentry.withProfiler(MapsIndoorsMap, { name: "MapsIndoorsMap" });