+++ packages/components/CHANGELOG.md
@@ -5,13 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [13.26.0] - 2024-02-20
+## [13.26.2] - 2025-05-08
+
+### Changed
+
+- Removed floor selector and my-position margin
+
+## [13.26.1] - 2025-04-29
+
+### Changed
+
+- Improved filtering in mi-dropdown component.
+
+## [13.26.0] - 2025-02-20
 
 ### Changed
 
 - Removed code responsible for `You have arrived.` last destination step.
 
-## [13.25.0] - 2024-01-29
+## [13.25.0] - 2025-01-29
 
 ### Added